requirements.txt
uvicorn==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
loguru==0.7.2
tenacity==8.5.0
aiohttp==3.10.9
greenlet==3.1.1
qdrant-client==1.12.0
langchain_text_splitters==0.3.2
langchain_core==0.3.15
langchain==0.3.7
fastapi==0.115.0
mistralai==1.2.5
sentence_transformers==3.3.1